chore(refactor): Split the router out into smaller logical units
