+++ packages/router/src/router.tsx
@@ -1,25 +1,21 @@
-import type { ReactNode, ReactElement } from 'react'
+import type { ReactNode } from 'react'
 import React, { useMemo, memo } from 'react'
 
 import { ActiveRouteLoader } from './active-route-loader'
 import { AuthenticatedRoute } from './AuthenticatedRoute'
 import { LocationProvider, useLocation } from './location'
+import { normalizePage } from './page'
 import { PageLoadingContextProvider } from './PageLoadingContext'
 import { ParamsProvider } from './params'
 import { Redirect } from './redirect'
-import type {
-  NotFoundRouteProps,
-  RedirectRouteProps,
-  RenderMode,
-} from './route-validators'
-import { isValidRoute, PageType } from './route-validators'
+import { Route } from './Route'
+import { isValidRoute } from './route-validators'
 import type { RouterContextProviderProps } from './router-context'
 import { RouterContextProvider } from './router-context'
 import { SplashPage } from './splash-page'
 import {
   analyzeRoutes,
   matchPath,
-  normalizePage,
   parseSearch,
   replaceParams,
   validatePath,
@@ -33,27 +29,6 @@ import type { AvailableRoutes } from './index'
 // Has to be `const`, or there'll be a race condition with imports in users'
 // projects
 const namedRoutes: AvailableRoutes = {}
-
-export interface RouteProps {
-  path: string
-  page: PageType
-  name: string
-  prerender?: boolean
-  renderMode?: RenderMode
-  whileLoadingPage?: () => ReactElement | null
-}
-
-/**
- * Route is now a "virtual" component
- * it is actually never rendered. All the page loading logic happens in active-route-loader
- * and all the validation happens within utility functions called from the Router
- */
-function Route(props: RouteProps): JSX.Element
-function Route(props: RedirectRouteProps): JSX.Element
-function Route(props: NotFoundRouteProps): JSX.Element
-function Route(_props: RouteProps | RedirectRouteProps | NotFoundRouteProps) {
-  return <></>
-}
 
 export interface RouterProps
   extends Omit<RouterContextProviderProps, 'routes' | 'activeRouteName'> {
@@ -292,10 +267,4 @@ const WrappedPage = memo(({ routeLoaderElement, sets }: WrappedPageProps) => {
   }, routeLoaderElement)
 })
 
-export {
-  Router,
-  Route,
-  namedRoutes as routes,
-  isValidRoute as isRoute,
-  PageType,
-}
+export { Router, Route, namedRoutes as routes, isValidRoute as isRoute }
