+++ packages/router/src/util.ts
@@ -2,13 +2,13 @@
 import type { ReactElement, ReactNode } from 'react'
 import { Children, isValidElement } from 'react'
 
+import type { PageType } from './page'
 import {
   isNotFoundRoute,
   isRedirectRoute,
   isStandardRoute,
   isValidRoute,
 } from './route-validators'
-import type { PageType } from './router'
 import { isPrivateNode, isPrivateSetNode, isSetNode } from './Set'
 
 export function flattenAll(children: ReactNode): ReactNode[] {
@@ -369,54 +369,6 @@ export function flattenSearchParams(queryString: string) {
   }
 
   return searchParams
-}
-
-export interface Spec {
-  name: string
-  prerenderLoader: (name?: string) => { default: React.ComponentType<unknown> }
-  LazyComponent:
-    | React.LazyExoticComponent<React.ComponentType<unknown>>
-    | React.ComponentType<unknown>
-}
-
-export function isSpec(
-  specOrPage: Spec | React.ComponentType,
-): specOrPage is Spec {
-  return (specOrPage as Spec).LazyComponent !== undefined
-}
-
-/**
- * Pages can be imported automatically or manually. Automatic imports are actually
- * objects and take the following form (which we call a 'spec'):
- *
- *   const WhateverPage = {
- *     name: 'WhateverPage',
- *     LazyComponent: lazy(() => import('src/pages/WhateverPage'))
- *     prerenderLoader: ...
- *   }
- *
- * Manual imports simply load the page:
- *
- *   import WhateverPage from 'src/pages/WhateverPage'
- *
- * Before passing a "page" to the PageLoader, we will normalize the manually
- * imported version into a spec.
- */
-export function normalizePage(
-  specOrPage: Spec | React.ComponentType<unknown>,
-): Spec {
-  if (isSpec(specOrPage)) {
-    // Already a spec, just return it.
-    return specOrPage
-  }
-
-  // Wrap the Page in a fresh spec, and put it in a promise to emulate
-  // an async module import.
-  return {
-    name: specOrPage.name,
-    prerenderLoader: () => ({ default: specOrPage }),
-    LazyComponent: specOrPage,
-  }
 }
 
 /**
